requirements.txt
# Python Dependencies

# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
openai-whisper>=20230314

# Diarization
pyannote.audio>=3.1.0

# Supabase
supabase>=2.0.0
postgrest-py>=0.13.0
pgvector>=0.2.0

# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# LiveKit (optional, for live ingestion)
livekit>=0.10.0
livekit-api>=0.5.0

# Evaluation & Metrics
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0
